ci: update and pin the actions used in generate workflow
This PR updates our github action versions to their latest, and pins
them. This is an action with elevated permissions, it has a write token
for pushing and PR creation.

* actions/checkout goes from 4 to 6.0.1
* actions/setup-node goes from 4 to 6.0.0
* peter-evans/create-pull-request goes from 6 to 7.0.9

I also enabled commit signing for the bot commit, which we should get
for free in the update.

(I guess I get coauthor on these automated PRs btw
bc I was the one who set up the cron trigger so it uses me as the actor?)